python-tutorial/postgressql_sample/requirements.txt
psycopg2-binary=2.9.6